plugin/file: Add RRResolver to Zone
Adds an RRResolver function to the Zone struct that when called reduces
the given slice of RRs down to a single RR, default is to return the
first RR in the slice.

When resolving CNAME records in externalLookup, the RRResolver function
is called instead of always assuming we want to use the first element in
the query, allowing any plugins using it to decide which element should
be used in teh case of multiple options in the zone cache.

Signed-off-by: Michael Nairn <mnairn@redhat.com>